pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genai-platform"
version = "0.1.0"
description = "A platform for building production-ready GenAI applications"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.80.0",
    "grpcio-tools>=1.80.0",
    "protobuf>=4.25.0",
    "httpx>=0.27.0",
    "mcp>=1.0",
    # Model Service provider SDKs. The corresponding ModelProvider /
    # EmbeddingProvider modules import these at module-import time, so
    # every install path (local dev, docker compose images, CI) needs
    # them — not optional extras.
    "openai>=1.50.0",
    "anthropic>=0.40.0",
    # Chapter 8: SDK runtime server inside per-workflow containers (Listings
    # 8.3-8.12) and the FastAPI app used by the gateway-side /jobs proxy.
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    # Chapter 8 deploy CLI emits Kubernetes manifests as YAML (Listing 8.23
    # adjacent — Deployment / HorizontalPodAutoscaler / Service).
    "pyyaml>=6.0",
]

[project.scripts]
genai-platform = "genai_platform.cli.deploy:main"

# Runtime extras — opt-in via `uv sync --extra postgres`.
[project.optional-dependencies]
postgres = [
    "psycopg2-binary>=2.9.0",
]

# PEP 735 dev dependency group — `uv sync` installs this by default
# (skip with `uv sync --no-dev`).
[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.4.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = ["*_pb2.py", "*_pb2_grpc.py"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.setuptools]
packages = {find = {}}

[tool.setuptools.package-data]
"proto" = ["*.proto"]

